Hack to fix bug with select2
select2 gets confused when multiple options have the same value… Perhaps
this isn’t supported behaviour. While it gets the option value right,
the option text could end up the text for some other option that happens
to share the same value.

Anyway, we have the right information by the time we hit select2:select,
so we update the text and title at that point.