Scrap `name_en`; opulate `name_local` if not 'en'
I wasn’t very keen on having name and name_en, with name_en only
populated if name was non English. So instead, `name` is always English;
`name_local` is present if the language is not English. This seems a bit
more intuitive, even to a non-native English speaker, hopefully.